checkMetAdd

+++ src/main/java/com/deusleydev/api_desafio_meli/domain/Produto.java
@@ -22,5 +22,5 @@ public class Produto {
     private String nome;
 
     @Column(name = "Valor do Produto")
-    private Integer valor;
+    private Double valor;
 }
